bot: Add skeleton handling for coffeebot commands
init and ping are coffeebot's concerns.

The way collectives are stored are fundamentally broken. Collectives
should be accessed by a stream ID and a subject name. This provides
Collective isolation, and removes any sort of state issue, and it's
clear which collective people are interacting with.

This should be done tomorrow. I've got until Thursday evening to
finish this and test it.