prod: Add coffeebot service file
A small note on how I do deployments currently:

Nix can build from source tarballs very easily. So instead of
uploading everything on to pypi and using that, I send the source
along with the service declaration to a special folder called packages
on my NixOS server configuration. I hard link in ./dist/ the
appropriate service.nix and default.nix files. The rest are in the
tarball. I also build quite often, remembering to bump version
numbers (one of the things I'd like to do is have a version number
only appear in one place, so that Nix and setup.py read the same
thing.